leg tracker covariance based on velocity of tracker, and fix rgb color map
git-svn-id: https://svn.code.sf.net/p/personalrobots/code/pkg/trunk@11257 f5854215-dd47-0410-b2c4-cdd35faa7885

+++ vision/people/src/filter/rgb.h
@@ -1,7 +1,7 @@
 #ifndef __RGB__
 #define __RGB__
 
-const static float rgb[] = {
+const static int rgb[] = {
  (0x00 << 16) | (0x00 << 8) | (0x77 << 00) ,
  (0x00 << 16) | (0x00 << 8) | (0x7f << 00) ,
  (0x00 << 16) | (0x00 << 8) | (0x80 << 00) ,
